pyproject.toml
[project]
name = "danbooru_vandalism_watch"
version = "0.1.0"
description = "A bot that monitors vandalism in a danbooru instance and reports it on discord."
authors = [{ name = "nonamethanks", email = "hellafrickingepic@gmail.com" }]
requires-python = ">=3.13,<3.14"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "danboorutools @ git+https://github.com/nonamethanks/danboorutools.git",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.8"
mypy = "^1.15.0"

[tool.poetry]
package-mode = false

[tool.ruff]
target-version = "py311"
line-length = 140

[tool.mypy]
check_untyped_defs = true
enable_error_code = "redundant-expr,ignore-without-code"
disable_error_code = "import-untyped"
warn_unused_ignores = true
show_column_numbers = true
show_error_codes = true
strict_equality = true

[tool.poetry.scripts]
bot = "run_bot:main"
